@@ -11,25 +11,34 @@ homepage = "https://questionpy.org"
 version = "0.2.0"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.11"
 pydantic = "^2.4"
 polyfactory = "^2.7.2"
 
-[tool.poetry.dev-dependencies]
-pylint = "^2.17.0"
+[tool.poetry.group.dev.dependencies]
+
+[tool.poetry.group.test]
+optional = true
+
+[tool.poetry.group.test.dependencies]
 pytest = "^7.2.2"
 pytest-md = "^0.2.0"
-pylint-pytest = "^1.1.2"
-mypy = "^1.1.1"
-flake8 = "^6.0.0"
 coverage = { extras = ["toml"], version = "^7.2.1" }
-tox = "^4.4.6"
 
+[tool.poetry.group.linter]
+dependencies = { ruff = "^0.2.2" }
+optional = true
+
+[tool.poetry.group.type-checker]
+dependencies = { mypy = "^1.1.1" }
+optional = true
+
+[tool.ruff]
+extend = "ruff_defaults.toml"
 
 [tool.pytest.ini_options]
 # https://github.com/pytest-dev/pytest-asyncio#auto-mode
 asyncio_mode = "auto"
-
 
 # This section is read automatically by Coverage.py when its working directory is .
 # https://coverage.readthedocs.io/en/6.5.0/config.html#configuration-reference
@@ -37,82 +46,8 @@ asyncio_mode = "auto"
 branch = true
 source = ["questionpy_common"]
 
-[tool.pylint]
-
-[tool.pylint.MAIN]
-extension-pkg-allow-list = "pydantic"
-# pylint_pytest removes some false positives when checking tests
-load-plugins = "pylint_pytest"
-fail-on = ["E", "F"]
-fail-under = 8
-
-[tool.pylint.BASIC]
-# Allow these well-known names despite usual naming conventions
-good-names = ["i", "j", "k", "e", "ex", "_"]
-# Allow any name for test methods
-function-rgx = "^test_.+|[a-z0-9_]+$"
-
-[tool.pylint.CLASSES]
-valid-metaclass-classmethod-first-arg = "mcs"
-
-[tool.pylint.DESIGN]
-max-attributes = 15
-max-branches = 15
-
-[tool.pylint.REPORTS]
-output-format = "colorized"
-
-[tool.pylint."MESSAGES CONTROL"]
-disable = [
-    "missing-module-docstring", "missing-class-docstring", "missing-function-docstring",
-    "unused-wildcard-import", "too-few-public-methods",
-]
-enable = ""
-
-[tool.pylint.FORMAT]
-expected-line-ending-format = "LF"
-indent-after-paren = 4
-indent-string = "    "
-max-line-length = 120
-
-[tool.pylint.IMPORTS]
-allow-wildcard-with-all = true
-
 [tool.mypy]
 plugins = "pydantic.mypy"
 disallow_untyped_defs = true
 strict_optional = true
 show_error_codes = true
-
-
-[tool.tox]
-legacy_tox_ini = """
-[tox]
-env_list =
-    py{39,310,311}-pytest
-    pylint
-    flake8
-    mypy
-
-minversion = 4.2.6
-
-[testenv]
-allowlist_externals = poetry
-# tox already installs the package (und thus its main dependencies) for us, so we only install dev dependencies here.
-commands_pre =
-    poetry install --no-root --without main
-commands =
-    coverage run -m pytest {tty:--color=yes} tests {posargs}
-
-[testenv:pylint]
-commands =
-    pylint questionpy_common tests
-
-[testenv:flake8]
-commands =
-    flake8 questionpy_common tests
-
-[testenv:mypy]
-commands =
-    mypy questionpy_common tests
-"""
