@@ -3,13 +3,23 @@
 #  (c) Technische Universität Berlin, innoCampus <info@isis.tu-berlin.de>
 
 from abc import ABC, abstractmethod
+from collections.abc import Sequence
 from enum import Enum
-from typing import Optional, Sequence, Annotated
+from typing import Annotated
 
 from pydantic import BaseModel, Field
 
-__all__ = ["CacheControl", "UiFile", "AttemptUi", "AttemptModel", "ScoringCode", "ClassifiedResponse",
-           "ScoreModel", "AttemptScoredModel", "BaseAttempt"]
+__all__ = [
+    "AttemptModel",
+    "AttemptScoredModel",
+    "AttemptUi",
+    "BaseAttempt",
+    "CacheControl",
+    "ClassifiedResponse",
+    "ScoreModel",
+    "ScoringCode",
+    "UiFile",
+]
 
 
 class CacheControl(Enum):
@@ -21,7 +31,7 @@ class CacheControl(Enum):
 class UiFile(BaseModel):
     name: str
     data: str
-    mime_type: Optional[str] = None
+    mime_type: str | None = None
 
 
 class AttemptUi(BaseModel):
@@ -29,8 +39,8 @@ class AttemptUi(BaseModel):
     """X(H)ML markup of the question UI."""
     placeholders: dict[str, str] = {}
     """Names and values of the ``<?p`` placeholders that appear in content."""
-    include_inline_css: Optional[str] = None
-    include_css_file: Optional[str] = None
+    include_inline_css: str | None = None
+    include_css_file: str | None = None
     cache_control: CacheControl = CacheControl.PRIVATE_CACHE
     files: list[UiFile] = []
 
@@ -41,10 +51,10 @@ class AttemptModel(BaseModel):
 
 
 class ScoringCode(Enum):
-    AUTOMATICALLY_SCORED = 'AUTOMATICALLY_SCORED'
-    NEEDS_MANUAL_SCORING = 'NEEDS_MANUAL_SCORING'
-    RESPONSE_NOT_SCORABLE = 'RESPONSE_NOT_SCORABLE'
-    INVALID_RESPONSE = 'INVALID_RESPONSE'
+    AUTOMATICALLY_SCORED = "AUTOMATICALLY_SCORED"
+    NEEDS_MANUAL_SCORING = "NEEDS_MANUAL_SCORING"
+    RESPONSE_NOT_SCORABLE = "RESPONSE_NOT_SCORABLE"
+    INVALID_RESPONSE = "INVALID_RESPONSE"
 
 
 class ClassifiedResponse(BaseModel):
@@ -57,9 +67,9 @@ class ClassifiedResponse(BaseModel):
 class ScoreModel(BaseModel):
     scoring_state: str = "{}"
     scoring_code: ScoringCode
-    score: Optional[float]
+    score: float | None
     """The total score for this question attempt, as a fraction of the default mark set by the LMS."""
-    classification: Optional[Sequence[ClassifiedResponse]] = None
+    classification: Sequence[ClassifiedResponse] | None = None
 
 
 class AttemptScoredModel(AttemptModel, ScoreModel):
@@ -67,7 +77,6 @@ class AttemptScoredModel(AttemptModel, ScoreModel):
 
 
 class BaseAttempt(ABC):
-
     @abstractmethod
     def export_attempt_state(self) -> str:
         """Serialize this attempt's relevant data.
