@@ -2,16 +2,29 @@
 #  QuestionPy is free software released under terms of the MIT license. See LICENSE.md.
 #  (c) Technische Universität Berlin, innoCampus <info@isis.tu-berlin.de>
 
-from typing import List, Literal, Union, Optional, get_args, Annotated
+from typing import Annotated, Literal, TypeAlias, TypeGuard, get_args
 
-from pydantic import BaseModel, PositiveInt, Field
-from typing_extensions import TypeGuard, TypeAlias
+from pydantic import BaseModel, Field, PositiveInt
 
 from questionpy_common.conditions import Condition
 
-__all__ = ["CanHaveConditions", "StaticTextElement", "TextInputElement", "CheckboxElement", "CheckboxGroupElement",
-           "Option", "RadioGroupElement", "SelectElement", "HiddenElement", "GroupElement", "RepetitionElement",
-           "FormElement", "FormSection", "OptionsFormDefinition", "is_form_element"]
+__all__ = [
+    "CanHaveConditions",
+    "CheckboxElement",
+    "CheckboxGroupElement",
+    "FormElement",
+    "FormSection",
+    "GroupElement",
+    "HiddenElement",
+    "Option",
+    "OptionsFormDefinition",
+    "RadioGroupElement",
+    "RepetitionElement",
+    "SelectElement",
+    "StaticTextElement",
+    "TextInputElement",
+    "is_form_element",
+]
 
 
 class _BaseElement(BaseModel):
@@ -28,6 +41,7 @@ class _Labelled(BaseModel):
 
 class CanHaveConditions(BaseModel):
     """Mixin class for elements that can have conditions on other elements."""
+
     disable_if: list[Condition] = []
     """Disable this element if any of these conditions match."""
     hide_if: list[Condition] = []
@@ -36,12 +50,14 @@ class CanHaveConditions(BaseModel):
 
 class CanHaveHelp(BaseModel):
     """Mixin class for elements that can have a help text hidden behind a button."""
-    help: Optional[str] = None
+
+    help: str | None = None
     """Text to be shown when the help button is clicked."""
 
 
 class StaticTextElement(_BaseElement, _Labelled, CanHaveConditions, CanHaveHelp):
     """Some static text with a label."""
+
     kind: Literal["static_text"] = "static_text"
     text: str
 
@@ -50,17 +66,17 @@ class TextInputElement(_BaseElement, _Labelled, CanHaveConditions, CanHaveHelp):
     kind: Literal["input"] = "input"
     required: bool = False
     """Require some non-empty input to be entered before the form can be submitted."""
-    default: Optional[str] = None
+    default: str | None = None
     """Default value of the input when first loading the form. Part of the submitted form data."""
-    placeholder: Optional[str] = None
+    placeholder: str | None = None
     """Placeholder to show when no value has been entered yet. Not part of the submitted form data."""
 
 
 class CheckboxElement(_BaseElement, CanHaveConditions, CanHaveHelp):
     kind: Literal["checkbox"] = "checkbox"
-    left_label: Optional[str] = None
+    left_label: str | None = None
     """Label shown the same way as labels on other element types."""
-    right_label: Optional[str] = None
+    right_label: str | None = None
     """Additional label shown to the right of the checkbox."""
     required: bool = False
     """Require this checkbox to be selected before the form can be submitted."""
@@ -70,12 +86,14 @@ class CheckboxElement(_BaseElement, CanHaveConditions, CanHaveHelp):
 
 class CheckboxGroupElement(_BaseElement):
     """Adds a 'Select all/none' button after multiple checkboxes."""
+
     kind: Literal["checkbox_group"] = "checkbox_group"
-    checkboxes: List[CheckboxElement]
+    checkboxes: list[CheckboxElement]
 
 
 class Option(BaseModel):
     """A possible option for radio groups and drop-downs."""
+
     label: str
     """Text describing the option, shown verbatim."""
     value: str
@@ -86,8 +104,9 @@ class Option(BaseModel):
 
 class RadioGroupElement(_BaseElement, _Labelled, CanHaveConditions, CanHaveHelp):
     """Group of radio buttons, of which at most one can be selected at a time."""
+
     kind: Literal["radio_group"] = "radio_group"
-    options: List[Option]
+    options: list[Option]
     """Selectable options."""
     required: bool = False
     """Require one of the options to be selected before the form can be submitted."""
@@ -95,10 +114,11 @@ class RadioGroupElement(_BaseElement, _Labelled, CanHaveConditions, CanHaveHelp)
 
 class SelectElement(_BaseElement, _Labelled, CanHaveConditions, CanHaveHelp):
     """A drop-down list."""
+
     kind: Literal["select"] = "select"
     multiple: bool = False
     """Allow the selection of multiple options."""
-    options: List[Option]
+    options: list[Option]
     """Selectable options."""
     required: bool = False
     """Require at least one of the options to be selected before the form can be submitted."""
@@ -106,18 +126,21 @@ class SelectElement(_BaseElement, _Labelled, CanHaveConditions, CanHaveHelp):
 
 class HiddenElement(_BaseElement, CanHaveConditions):
     """An element that isn't shown to the user but still submits its fixed value."""
+
     kind: Literal["hidden"] = "hidden"
     value: str
 
 
 class GroupElement(_BaseElement, _Labelled, CanHaveConditions, CanHaveHelp):
     """Groups multiple elements horizontally with a common label."""
+
     kind: Literal["group"] = "group"
-    elements: List["FormElement"]
+    elements: list["FormElement"]
 
 
 class RepetitionElement(_BaseElement):
     """Repeats a number of elements, allowing the user to add new repetitions with the click of a button."""
+
     kind: Literal["repetition"] = "repetition"
 
     initial_repetitions: PositiveInt
@@ -126,33 +149,42 @@ class RepetitionElement(_BaseElement):
     """Minimum number of repetitions, at or below which removal is not possible."""
     increment: PositiveInt
     """Number of repetitions to add with each click of the button."""
-    button_label: Optional[str] = None
+    button_label: str | None = None
     """Label for the button that adds more repetitions, or None to use default provided by LMS."""
 
     elements: list["FormElement"]
     """Elements that will be repeated."""
 
 
-FormElement: TypeAlias = Annotated[Union[
-    StaticTextElement, TextInputElement, CheckboxElement, CheckboxGroupElement,
-    RadioGroupElement, SelectElement, HiddenElement, GroupElement, RepetitionElement
-], Field(discriminator="kind")]
+FormElement: TypeAlias = Annotated[
+    CheckboxElement
+    | CheckboxGroupElement
+    | GroupElement
+    | HiddenElement
+    | RadioGroupElement
+    | RepetitionElement
+    | SelectElement
+    | StaticTextElement
+    | TextInputElement,
+    Field(discriminator="kind"),
+]
 
 
 class FormSection(BaseModel):
     """Form section that can be expanded and collapsed."""
+
     name: str
     """Name that will later identify the element in submitted form data."""
     header: str
     """Header to be shown at the top of the section."""
-    elements: List[FormElement] = []
+    elements: list[FormElement] = []
     """Elements contained in the section."""
 
 
 class OptionsFormDefinition(BaseModel):
-    general: List[FormElement] = []
+    general: list[FormElement] = []
     """Elements to add to the main section, after the LMS' own elements."""
-    sections: List[FormSection] = []
+    sections: list[FormSection] = []
     """Sections to add after the main section."""
 
 
